Fix bad imports from server > client
