+++ packages/locales/package.json
@@ -9,7 +9,8 @@
     "./client": "./src/client.ts",
     "./provider": "./src/provider.tsx",
     "./lang": "./src/lang.ts",
-    "./translation-getters": "./src/translation-getters.ts"
+    "./hooks": "./src/hooks.ts",
+    "./server-hooks": "./src/server-hooks.ts"
   },
   "scripts": {
     "clean": "rm -rf .turbo node_modules",
